Properly moving remoteOpPreffered to runner client
It makes more sense tied to the runner client - the runner is really what cares if the operation will happen locally or remotely - it needs to know if it should spool up a local runner.
